Remvoe last few v2 directories
